Refactored operation factory to seperate module
